+++ zellij-server/src/panes/floating_panes/mod.rs
@@ -406,7 +406,7 @@ impl FloatingPanes {
                     output.add_pane_contents(
                         &connected_clients,
                         pane.pid(),
-                        pane.pane_contents(None, false),
+                        pane.pane_contents(None, false, None),
                     );
                 },
                 PaneId::Plugin(_) => {
@@ -414,7 +414,7 @@ impl FloatingPanes {
                         output.add_pane_contents(
                             &[*client_id],
                             pane.pid(),
-                            pane.pane_contents(Some(*client_id), false),
+                            pane.pane_contents(Some(*client_id), false, None),
                         );
                     }
                 },