+++ zellij-utils/assets/prost/api.plugin_command.rs
@@ -1860,6 +1860,8 @@ pub struct RegexHighlight {
     pub italic: bool,
     #[prost(bool, tag="7")]
     pub underline: bool,
+    #[prost(string, optional, tag="8")]
+    pub tooltip_text: ::core::option::Option<::prost::alloc::string::String>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]