+++ zellij-utils/src/vendored/termwiz/mod.rs
@@ -1,0 +1,6 @@
+// Thid module was inlined from the termwiz library: https://github.com/wezterm/wezterm/tree/main/termwiz
+//
+// Most of it was stubbed out and some parts have been adjusted to fit
+pub mod input;
+mod keymap;
+mod readbuf;